static roadmap items on ai page
